Rename package from com.android.sample to com.swentseekr.seekr
Updated all source, test, and resource files to use the new package name com.swentseekr.seekr. Changed applicationId, namespace, and manifest package, updated references in code imports, and renamed the app from SampleApp to Seekr in strings and settings.
