+++ app/google-services.json
@@ -9,7 +9,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:1040243223335:android:90a92630d4c0d6e2e79702",
         "android_client_info": {
-          "package_name": "com.android.sample"
+          "package_name": "com.swentseekr.seekr"
         }
       },
       "oauth_client": [],
