+++ app/src/main/java/com/swentseekr/seekr/model/author/Author.kt
@@ -1,4 +1,4 @@
-package com.android.sample.model.author
+package com.swentseekr.seekr.model.author
 
 data class Author(
     val pseudonym: String,
