+++ app/src/main/java/com/swentseekr/seekr/model/map/Location.kt
@@ -1,3 +1,3 @@
-package com.android.sample.model.map
+package com.swentseekr.seekr.model.map
 
 data class Location(val latitude: Double, val longitude: Double, val name: String)
