+++ app/src/test/java/com/swentseekr/seekr/ExampleRobolectricTest.kt
@@ -1,9 +1,9 @@
-package com.android.sample
+package com.swentseekr.seekr
 
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.android.sample.screen.SecondScreen
 import com.kaspersky.kaspresso.testcases.api.testcase.TestCase
+import com.swentseekr.seekr.screen.SecondScreen
 import io.github.kakaocup.compose.node.element.ComposeScreen
 import org.junit.Rule
 import org.junit.Test
