+++ app/src/test/java/com/swentseekr/seekr/PointTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.swentseekr.seekr
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
